Fix: Fix bus route solution

+++ src/Competitions/KickStart2020/BusRoutes.java
@@ -9,16 +9,16 @@ public class BusRoutes {
         int T = input.nextInt();
         for (int ks = 1; ks <= T; ks++) {
             int noBuses = input.nextInt();
-            int lastDay = input.nextInt();
-            int[] buses = new int[noBuses];
-            for (int i = 0; i < noBuses; i++) buses[i] = input.nextInt();
+            long lastDay = input.nextLong();
+            long[] buses = new long[noBuses];
+            for (int i = 0; i < noBuses; i++) buses[i] = input.nextLong();
             System.out.println(String.format("Case #%d: %d", ks, solve(lastDay, buses)));
         }
     }
 
-    private static int solve(int lastDay, int[] buses) {
+    private static long solve(long lastDay, long[] buses) {
         for (int i = buses.length - 1; i >= 0; i--) {
-            int bus = buses[i];
+            long bus = buses[i];
             lastDay = Math.min(bus * (lastDay / bus), lastDay);
         }
         return lastDay;
